chat-app/backend/requirements.txt
litellm>=1.40.14
fastapi>=0.110.0
uvicorn>=0.29.0
python-dotenv>=1.0.0
pydantic>=2.0.0
openai>=1.0.0
